requirements.txt
# ArXiv RAG Assistant - Dependencies
# Install: conda run -n pytorch pip install -r requirements.txt

# Embeddings & ML
sentence-transformers>=2.2.0
chromadb>=0.4.0
rank_bm25>=0.2.2
tiktoken>=0.5.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
groq>=0.9.0

# ArXiv ingestion
requests>=2.31.0
lxml>=4.9.0
feedparser>=6.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
